Update when Testrunner moves core plugin to coreplugin module
