+++ pom.xml
@@ -67,7 +67,7 @@
     <dependencies>
         <dependency>
             <groupId>edu.illinois.cs</groupId>
-            <artifactId>testrunner-maven-plugin</artifactId>
+            <artifactId>testrunner-core-plugin</artifactId>
             <version>1.1</version>
         </dependency>
